\documentclass{article} \usepackage[utf8]{inputenc}  \title{Gödel's incompleteness theorems} \author{Submitted By Gayathri  sadagopan} \date{22nd July 2021}  \begin{document}  \maketitle  \section{Introduction} Gödel's incompleteness theorems are two theorems of mathematical logic that are concerned with the limits of probability in formal axiomatic theories. The theorems are widely used to find a complete and consistent set of axioms for all mathematics is impossible. The incompleteness theorems apply to formal systems.
Formal system is a deductive apparatus that consists of a particular set of axioms along with rules of symbolic manipulation (or rules of inference) that allow for the derivation of new theorems from the axioms. There are several properties that a formal system may have, including completeness, consistent, and the existence of an effective axiomatization. The incompleteness theorems show that systems which contain a sufficient amount of arithmetic cannot possess all three of these properties.It has two type of incompleteness theorem that is

A) First incompleteness theorem

First Incompleteness Theorem is "Any consistent formal system F within which a certain amount of elementary arithmetic can be carried out is incomplete; i.e., there are statements of the language of F which can neither be proved nor disproved in F." 

B) Second incompleteness theorem 

Second Incompleteness Theorem is "Assume F is a consistent formalized system which contains elementary arithmetic.

This theorem is stronger than the first incompleteness theorem because the statement constructed in the first incompleteness theorem does not directly express the consistency of the system. The proof of the second incompleteness theorem is obtained by formalizing the proof of the first incompleteness theorem within the system F itself.

\section{Relationship with computability}
The incompleteness theorem is closely related to several results about undecidable sets in recursion theory. Kleene presented a proof of Gödel's incompleteness theorem using basic results of computability theory. One such result shows that the halting problem is undecidable such as there is no computer program that can correctly determine, given any program P as input, whether P eventually halts when run with a particular given input. Kleene showed that the existence of a complete effective system of arithmetic with certain consistency properties would force the halting problem to be decidable, a contradiction.

The incompleteness results affect the philosophy of mathematics, particularly versions of formalism, which use a single system of formal logic to define their principles. The incompleteness theorem is sometimes thought to have severe consequences for the program of localism. Gödel's incompleteness theorems imply about human intelligence Much of the debate centers on whether the human mind is equivalent to a Turing machine. Or by the Church–Turing thesis, any finite machine at all. If it is, and if the machine is consistent, then Gödel's incompleteness theorems would apply to it.
\end{document}